refactoring Product +create discount class
