--- CalculatorTask/Price.cs
@@ -1,6 +1,0 @@
-﻿namespace CalculatorTask
-{
-    internal class Price
-    {
-    }
-}
