--- CalculatorTask/ProductDiscount.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CalculatorTask
-{
-    internal class ProductDiscount
-    {
-        private double DiscountPercentage;
-        private double price;
-
-
-
-    }
-}
